Finished installations manual and tools

+++ tools/importIntoNAS.sh
@@ -1,14 +1,41 @@
-
+ME=`basename $0`
 # arg 1: seed or file
-# needs webdanica-core, common-core-5.1, harvester-core-5.1, JDBC driver for NAS harvestdatabase
+# needs webdanica-core, common-core-5.2.2, harvester-core-5.2.2, JDBC driver for NAS harvestdatabase
 SETTINGSFILE=conf/settings_WebdanicaClient.xml
+if [ ! -f "$SETTINGSFILE" ]; then
+ echo ERROR: The Netarchivesuite settingsfile \"$SETTINGSFILE\" does not exist. Please correct the path in $ME
+ exit
+fi
+
 SILENTLOGBACKFILE=conf/silent_logback.xml
 OPTS1=-Ddk.netarkivet.settings.file=$SETTINGSFILE
 OPTS2=-Dlogback.configurationFile=$SILENTLOGBACKFILE
-WEBDANICAJAR=lib/webdanica-core-1.0.0.jar
+
+NAS_VERSION=5.2.2
+VERSION=2.0
+
+WEBDANICA_JAR=lib/webdanica-core-$VERSION.jar
+if [ ! -f "$WEBDANICA_JAR" ]; then
+ echo ERROR: The webdanica jarfile \"$WEBDANICA_JAR\" does not exist. Please correct the path in this script \($ME\)
+ exit
+fi
+
 DATABASEJAR=lib/postgresql-9.2-1003-jdbc4.jar
-NAS_VERSION=5.1
+if [ ! -f "$DATABASEJAR" ]; then
+ echo ERROR: The database jarfile \"$DATABASEJAR\" does not exist. Please correct the path in this script \($ME\)
+ exit
+fi
+
 NAS_COMMON_JAR=lib/common-core-$NAS_VERSION.jar
+if [ ! -f "$NAS_COMMON_JAR" ]; then
+ echo ERROR: The Netarchivesuite common-core jarfile \"$NAS_COMMON_JAR\" does not exist. Please correct the path in this script \($ME\)
+ exit
+fi
 NAS_HARVESTER_JAR=lib/harvester-core-$NAS_VERSION.jar
+if [ ! -f "$NAS_HARVESTER_JAR" ]; then
+ echo ERROR: The Netarchivesuite harvester-core \"$NAS_HARVESTER_JAR\" does not exist. Please correct the path in this script \($ME\)
+ exit
+fi
+
 
 java $OPTS1 $OPTS2 -cp $WEBDANICAJAR:$DATABASEJAR:$NAS_COMMON_JAR:$NAS_HARVESTER_JAR dk.kb.webdanica.core.tools.ImportIntoNetarchiveSuite $1
